backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
distro==1.9.0
docopt==0.6.2
fastapi==0.135.1
groq==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Js2Py==0.74
packaging==26.0
pipwin==0.5.2
pydantic==2.12.5
pydantic_core==2.41.5
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
